+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [6.2.0] - 2019-04-22
+### Fixed
+- Proxy connection pool key for activities and telemetry
+
+### Added
+- Enforcer telemetry by request
+
 ## [6.1.1] - 2019-04-16
 ### Fixed
 - Proxy connection pool and scheme handling